format_code.sh: Fix failure if files were removed
(cherry picked from commit ca913760fe698340420a623f846d55f4bfb28617)

+++ dev/format_code.sh
@@ -62,18 +62,20 @@ do
   UNTRACKED=$(git ls-files --others --exclude-standard -- "${dir}**.cpp" "${dir}**.hpp" "${dir}**.h")
   for file in $TRACKED $UNTRACKED
   do
-    # Note: Do NOT use in-place edition of clang-format because this causes
-    # "make" to detect the files as changed every time, even if the content was
-    # not modified! So we only overwrite the files if their content has changed.
-    OLD_CONTENT=$(cat "$file")
-    NEW_CONTENT=$(clang-format -style=file "$file")
-    if [ "$NEW_CONTENT" != "$OLD_CONTENT" ]
-    then
-        printf "%s\n" "$NEW_CONTENT" > "$file"
-        echo "[M] $file"
-        COUNTER=$((COUNTER+1))
-    else
-        echo "[ ] $file"
+    if [ -f "$file" ]; then
+      # Note: Do NOT use in-place edition of clang-format because this causes
+      # "make" to detect the files as changed every time, even if the content was
+      # not modified! So we only overwrite the files if their content has changed.
+      OLD_CONTENT=$(cat "$file")
+      NEW_CONTENT=$(clang-format -style=file "$file")
+      if [ "$NEW_CONTENT" != "$OLD_CONTENT" ]
+      then
+          printf "%s\n" "$NEW_CONTENT" > "$file"
+          echo "[M] $file"
+          COUNTER=$((COUNTER+1))
+      else
+          echo "[ ] $file"
+      fi
     fi
   done
 done
